+++ apps/scrybe/feature/settings/src/main/kotlin/dev/scrybe/feature/settings/AIConfigScreen.kt
@@ -62,7 +62,7 @@ import com.twobits.design.components.LocalModelPanel
 import com.twobits.design.components.LocalModelStatus
 import com.twobits.design.components.ModelRadioList
 import dev.scrybe.core.common.ScrybeLayoutDefaults
-import dev.scrybe.core.model.LocalLlmModel
+import com.twobits.core.localmodels.LocalLlmModel
 import dev.scrybe.core.model.LocalWhisperModel
 import dev.scrybe.core.model.OpenAiTranscriptionModel
 import dev.scrybe.core.model.OpenAiTransformModel